Typo fix for assigning secondary structure
The secondary structure (ss) assignment had a typo that caused problems when the ss was assigned to multi chain models with different chain lengths. When chain.set_ss is run it must either have a input of length one or input length greater than or equal to the length of the chain.